+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### WIP
+
+  - Added UIkit.Utils.throttle helper function
+
 ### 2.26.2 (April 06, 2016)
   - Added iframe support for lightbox component
   - Added observe option to data-uk-grid-margin + data-uk-grid-match